# connectease_backend/openapi.yaml
openapi: 3.0.3
info:
  title: ConnectEase API
  description: >
    REST API for ConnectEase — a service marketplace connecting customers with
    vendors for home services, PGs, food, electrical, plumbing, cleaning, and laundry.

    **Authentication:** All protected endpoints read a JWT from an HttpOnly cookie
    named `jwt`. Call `POST /api/auth/signin` first — the cookie is set automatically.

    **Roles:**
    - `customer` — can browse, review, chat, post to community
    - `vendor`   — same as customer plus full service listing management (`/api/vendor/**`)
  version: 1.0.0
  contact:
    name: ConnectEase Team
servers:
  - url: http://localhost:8081
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: >
        HttpOnly JWT cookie set by POST /api/auth/signin.
        In Swagger UI click "Authorize", paste the raw JWT value.
        In Postman the cookie is stored automatically after login.

# ─────────────────────────────────────────────────────────────────────────────
# SCHEMAS
# ─────────────────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Operation completed successfully
...
    UserProfileDTO:
      type: object
      properties:
        uid:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        fullName:
          type: string
          example: Ravi Kumar
        email:
          type: string
          example: ravi@example.com
        phoneNo:
          type: string
          nullable: true
          example: "9876543210"
        image:
          type: string
          nullable: true
          example: https://cdn.example.com/avatars/ravi.jpg
        role:
          type: string
          enum: [customer, vendor]
          example: customer
        createdAt:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00"
    UpdateProfileRequest:
      type: object
      description: All fields are optional. Only provided fields are updated.
      properties:
        fullName:
          type: string
          example: Ravi Kumar Updated
        phoneNo:
          type: string
          example: "9123456789"
        image:
          type: string
          example: https://cdn.example.com/avatars/ravi-new.jpg
    ChangePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          example: ravi1234
        newPassword:
          type: string
          example: newSecurePass99

    # ── Services / Listings ──────────────────────────────────────────────────
    ImageDTO:
      type: object
      properties:
        url:
          type: string
          example: https://cdn.example.com/services/cleaning1.jpg
        isPrimary:
          type: boolean
          example: true
    ReviewDTO:
      type: object
      properties:
        userName:
          type: string
          example: Ravi Kumar
        profileImage:
          type: string
          nullable: true
          example: https://cdn.example.com/avatars/ravi.jpg
        review:
          type: string
          example: Excellent service, very thorough!
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
    ListingCardDTO:
      type: object
      properties:
        sid:
          type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Home Deep Cleaning
        description:
          type: string
          example: Professional full-home cleaning service
        price:
          type: number
          format: double
          example: 1500.00
        categoryName:
          type: string
          example: cleaning service
        city:
          type: string
          example: Hyderabad
        area:
          type: string
          example: Gachibowli
        primaryImageUrl:
          type: string
          nullable: true
          example: https://cdn.example.com/services/cleaning.jpg
    ServiceDetailsDTO:
      type: object
      properties:
        sid:
          type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Home Deep Cleaning
        description:
          type: string
          example: Full home deep cleaning including kitchen and bathrooms
        price:
          type: number
          format: double
          example: 1500.00
        totalViews:
          type: integer
          format: int64
          example: 243
        vendorName:
          type: string
          example: CleanPro Services
        averageRating:
          type: number
          format: double
          example: 4.3
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDTO'
        images:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ImageDTO'
    ReviewRequestDTO:
      type: object
      required:
        - score
      properties:
        review:
          type: string
          example: Great service, highly recommended!
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 5

    # ── Vendor ────────────────────────────────────────────────────────────────
...
    CommunityPostDTO:
      type: object
      properties:
        postId:
          type: string
          example: p1b2c3d4-e5f6-7890-abcd-ef1234567890
        title:
          type: string
          example: Best PG in Gachibowli?
        description:
          type: string
          example: Looking for recommendations for a good PG near Cyber Gateway
        image:
          type: string
          nullable: true
          example: https://cdn.example.com/posts/pg-search.jpg
        time:
          type: string
          format: date-time
          example: "2025-03-10T14:22:00"
        authorFullName:
          type: string
          example: Ravi Kumar
...
    ChatMessageDto:
      type: object
      properties:
        messageId:
          type: string
          example: m1a2b3c4-d5e6-7890-abcd-ef1234567890
        senderId:
          type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        senderName:
          type: string
          example: Ravi Kumar
        senderImage:
          type: string
          nullable: true
          example: null
        content:
          type: string
          example: Hi, is your service available this weekend?
        createdAt:
          type: string
          format: date-time
          example: "2025-03-10T15:30:00"
    ChatSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          example: cs1b2c3d-e5f6-7890-abcd-ef1234567890
        participantName:
          type: string
          example: CleanPro Services
        participantImage:
          type: string
          nullable: true
          example: https://cdn.example.com/avatars/cleanpro.jpg
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageDto'

    # ── AI Chat ───────────────────────────────────────────────────────────────
...
    PageListingCardDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ListingCardDTO'
        totalElements:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 3
        size:
          type: integer
          example: 10
        number:
          type: integer
          description: Current page index (0-based)
          example: 0
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
...
paths:
...
  /api/users/{uid}:
    get:
      tags: [Users]
      summary: Get user profile (public)
      operationId: getUserProfile
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      responses:
        '200':
          description: User profile (password is never returned)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
              example:
                uid: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                fullName: Ravi Kumar
                email: ravi@example.com
                phoneNo: "9876543210"
                image: https://cdn.example.com/avatars/ravi.jpg
                role: customer
                createdAt: "2025-01-15T10:30:00"
        '500':
          description: User not found

    put:
      tags: [Users]
      summary: Update own profile
      operationId: updateProfile
      security:
        - cookieAuth: []
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              fullName: Ravi Kumar Updated
              phoneNo: "9123456789"
              image: https://cdn.example.com/avatars/ravi-new.jpg
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
        '500':
          description: User not found or trying to update another user's profile

    delete:
      tags: [Users]
      summary: Delete own account
      operationId: deleteAccount
      security:
        - cookieAuth: []
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: success
                message: Account deleted successfully
        '500':
          description: User not found or trying to delete another user's account
  /api/users/{uid}/password:
    put:
      tags: [Users]
      summary: Change own password
      operationId: changePassword
      security:
        - cookieAuth: []
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              oldPassword: ravi1234
              newPassword: newSecurePass99
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: success
                message: Password updated successfully
        '500':
          description: Incorrect old password or unauthorized

  # ════════════════════════════════════════════════════════════════════════════
  # LISTINGS (FILTER)
  # ════════════════════════════════════════════════════════════════════════════
...
  /api/services:
    get:
      tags: [Services]
      summary: Browse all services (paginated, sorted newest first)
      operationId: getAllServices
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated service cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageListingCardDTO'
              example:
                content:
                  - sid: s1b2c3d4-e5f6-7890-abcd-ef1234567890
                    name: Home Deep Cleaning
                    description: Professional full-home cleaning
                    price: 1500.00
                    categoryName: cleaning service
                    city: Hyderabad
                    area: Gachibowli
                    primaryImageUrl: https://cdn.example.com/services/cleaning.jpg
                totalElements: 25
                totalPages: 3
                size: 10
                number: 0
                first: true
                last: false
  /api/services/{sid}:
    get:
      tags: [Services]
      summary: Get full service details (with reviews and images)
      operationId: getServiceDetails
      parameters:
        - name: sid
          in: path
          required: true
          schema:
            type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
      responses:
        '200':
          description: Service detail page data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailsDTO'
              example:
                sid: s1b2c3d4-e5f6-7890-abcd-ef1234567890
                name: Home Deep Cleaning
                description: Full home deep cleaning including kitchen and bathrooms
                price: 1500.00
                totalViews: 243
                vendorName: CleanPro Services
                averageRating: 4.3
                reviews:
                  - userName: Ravi Kumar
                    profileImage: null
                    review: Excellent service, very thorough!
                    score: 5
                  - userName: Priya Sharma
                    profileImage: https://cdn.example.com/avatars/priya.jpg
                    review: Good but slightly late
                    score: 4
                images:
                  - url: https://cdn.example.com/services/cleaning1.jpg
                    isPrimary: true
                  - url: https://cdn.example.com/services/cleaning2.jpg
                    isPrimary: false
        '500':
          description: Service not found
  /api/services/vendor/{vendorId}:
    get:
      tags: [Services]
      summary: Get all services listed by a specific vendor (public)
      operationId: getServicesByVendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
          example: b2c3d4e5-f6a7-8901-bcde-f12345678901
      responses:
        '200':
          description: List of vendor services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListingCardDTO'
              example:
                - sid: s1b2c3d4-e5f6-7890-abcd-ef1234567890
                  name: Home Deep Cleaning
                  description: Professional full-home cleaning
                  price: 1500.00
                  categoryName: cleaning service
                  city: Hyderabad
                  area: Gachibowli
                  primaryImageUrl: https://cdn.example.com/services/cleaning.jpg
  /api/services/{sid}/reviews:
    post:
      tags: [Services]
      summary: Add a review to a service
      operationId: addReview
      description: >
        The `userId` field is ignored even if provided.
        The server always uses the authenticated user's ID from the JWT cookie.
      security:
        - cookieAuth: []
      parameters:
        - name: sid
          in: path
          required: true
          schema:
            type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequestDTO'
            example:
              review: Great service, highly recommended!
              score: 5
      responses:
        '200':
          description: Review added
          content:
            text/plain:
              schema:
                type: string
              example: Review added successfully!
        '401':
          description: Unauthorized — must be logged in
        '500':
          description: Service not found

  # ════════════════════════════════════════════════════════════════════════════
  # VENDOR (VENDOR ROLE REQUIRED)
  # ════════════════════════════════════════════════════════════════════════════
...
  /api/community/user/{uid}:
    get:
      tags: [Community]
      summary: Get all posts by a specific user
      operationId: getPostsByUser
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
      responses:
        '200':
          description: User's community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommunityPostDTO'
              example:
                - postId: p1b2c3d4-e5f6-7890-abcd-ef1234567890
                  title: Best PG in Gachibowli?
                  description: Looking for recommendations
                  image: null
                  time: "2025-03-10T14:22:00"
                  authorFullName: Ravi Kumar

  # ════════════════════════════════════════════════════════════════════════════
  # CHAT
  # ════════════════════════════════════════════════════════════════════════════
  /api/chat/start/{vendorId}:
    get:
      tags: [Chat]
      summary: Open or initialize a chat session with a vendor
      operationId: startChat
      description: >
        Returns the existing session if one already exists between the logged-in
        user and the target vendor, otherwise creates a new session.
        Use the returned `sessionId` to connect via WebSocket.

        **WebSocket (STOMP over SockJS):**

        - Connect: `ws://localhost:8081/ws-chat`
        - Subscribe: `/topic/session/{sessionId}`
        - Send messages to: `/app/chat.sendMessage`
        - Message payload: `{ "sessionId": "...", "content": "Hello!" }`
      security:
        - cookieAuth: []
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: string
          example: b2c3d4e5-f6a7-8901-bcde-f12345678901
      responses:
        '200':
          description: Chat session with history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionResponse'
              example:
                sessionId: cs1b2c3d-e5f6-7890-abcd-ef1234567890
                participantName: CleanPro Services
                participantImage: https://cdn.example.com/avatars/cleanpro.jpg
                messages:
                  - messageId: m1a2b3c4-d5e6-7890-abcd-ef1234567890
                    senderId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                    senderName: Ravi Kumar
                    senderImage: null
                    content: Hi, is your service available this weekend?
                    createdAt: "2025-03-10T15:30:00"
        '401':
          description: Unauthorized — must be logged in

  # ════════════════════════════════════════════════════════════════════════════
  # AI CHAT
  # ════════════════════════════════════════════════════════════════════════════
...
  /api/locations/cities/{city}/areas:
    get:
      tags: [Locations]
      summary: Get all areas within a specific city
      operationId: getAreasByCity
      description: Useful for populating area filter dropdowns after a city is selected.
      parameters:
        - name: city
          in: path
          required: true
          schema:
            type: string
          example: Hyderabad
      responses:
        '200':
          description: Alphabetically sorted list of area names in the city
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example:
                - Ameerpet
                - Gachibowli
                - Kondapur
                - Kukatpally
                - Madhapur