# connectease_backend/src/main/java/com/cts/connectease/dto/ChatMessageDto.java
package com.cts.connectease.dto;

import lombok.Data;

@Data
public class ChatMessageDto {
    private String messageId;
    private String sessionId;
    private String senderId;
    private String senderName;
    private String senderImage;
    private String content;
    private String createdAt;
}